update font style
update font style
